requirements.txt
termcolor==1.1.0
requests==2.18.4
PixivPy==3.5.5
